改改bug

+++ f_db.php
@@ -8,8 +8,24 @@
      * @author	TangSong jim20259@gmail.com
      * @link	http://www.plurk.com/tangsong
      */
+    require_once('config.php');
+    $test = new database;
+    $test->getURL(2);
+
+    //$test->getUsername(2);
+
+    //$test->getSiteName(1);
+
+    $test->newUser('newuserh','1234','emmmaaail@email.com');
+
+    $test->addList(2,'www.google.com.tw','Google');
+
+    $test->addList(2,'www.yahoo.com.tw','yahoo');
+
+    $test->addList(2,'www.google.com.tw','gggggle');
+
+    $test->deleteURL('admin2','www.yahoo.com.tw');
     class database{
-        require_once('config.php');
         protected $db;
 
         public function __Construct(){
@@ -43,9 +59,10 @@
          * @return void
          */
         public function getURL($user_id){
-            $query="SELECT URL FROM user_url WHERE ID = ?";
-            $sth = $this->db->query($query);
-            $sth->execute( array($user_id) );
+            $query="SELECT URL FROM user_url WHERE ID = :user_id";
+            $sth = $this->db->prepare($query);
+            $sth->bindParam(':user_id', $user_id);
+            $sth->execute();
             $i=0;
             echo '<script>';
             foreach( $sth as $row ){
@@ -64,9 +81,10 @@
          * @return void
          */
         public function getUsername($user_id){
-            $query="SELECT username FROM user_url WHERE ID = ?";
+            $query="SELECT username FROM user_url WHERE ID = :user_id";
             $sth = $this->db->query($query);
-            $sth->execute( array($user_id) );
+            $sth->bindParam(':user_id', $user_id);
+            $sth->execute();
             $i=0;
             echo '<script>';
             foreach( $sth as $row ){
@@ -85,9 +103,10 @@
          * @return void
          */
         public function getSiteName($user_id){
-            $query="SELECT site_name FROM user_url WHERE ID = ?";
+            $query="SELECT site_name FROM user_url WHERE ID = :user_id";
             $sth = $this->db->query($query);
-            $sth->execute( array($user_id) );
+            $sth->bindParam(':user_id', $user_id);
+            $sth->execute();
             $i=0;
             echo '<script>';
             foreach( $sth as $row ){
@@ -112,9 +131,12 @@
                 return false;
             }else{
                 $password = md5($password);
-                $query = "INSERT INTO Account (username,password,email) VALUES (?,?,?)";
+                $query = "INSERT INTO Account (username,password,email) VALUES (:username, :password, :email)";
                 $sth = $this->db->prepare($query);
-                $sth->execute( array($username, $password, $email) );
+                $sth->bindParam(':username', $username);
+                $sth->bindParam(':password', $password);
+                $sth->bindParam(':email', $email);
+                $sth->execute();
                 return true;
             }
         }
@@ -126,9 +148,10 @@
          * @return boolean 帳號名稱是否存在
          */
         private function checkUser($username){
-            $query = "SELECT username FROM Account WHERE username =?";
+            $query = "SELECT username FROM Account WHERE username = :username";
             $sth = $this->db->prepare($query);
-            $sth->execute( array($username) );
+            $sth->bindParam(':username', $username);
+            $sth->execute();
             foreach($sth as $row){
                 if($row['username'] == $username ){
                     return true;
@@ -146,15 +169,18 @@
          * @param string $site_name 網站名稱
          * @return void
          */
-        private function addList($user_id,$URL,$site_name){
-            $checkURL = checkURL($URL);
+        public function addList($user_id,$URL,$site_name){
+            $checkURL = $this->checkURL($URL);
             if($checkURL){
-                $query = "INSERT INTO List (UID, URL_ID, site_name) VALUES (?,?,?)";
+                $query = "INSERT INTO List (UID, URL_ID, site_name) VALUES (:user_id, :URL, :site_name)";
                 $sth = $this->db->prepare($query);
-                $sth->execute( array($user_id, $checkURL, $site_name) );
+                $sth->bindParam(':user_id', $user_id);
+                $sth->bindParam(':URL', $URL);
+                $sth->bindParam(':site_name', $site_name);
+                $sth->execute();
             }else{
-                newURL($URL);
-                addList($user_id,$URL,$site_name);
+                $this->newURL($URL);
+                $this->addList($user_id,$URL,$site_name);
             }
         }
 
@@ -165,9 +191,11 @@
          * @return void
          */
         private function newURL($URL){
-            $query = "INSERT INTO URL (HASH, URL) VALUES (?,?)";
+            $query = "INSERT INTO URL (HASH, URL) VALUES (:HASH, :URL)";
             $sth = $this->db->prepare($query);
-            $sth->execute( array(md5($URL), $URL) );
+            $sth->bindParam(':HASH', md5($URL));
+            $sth->bindParam(':URL', $URL);
+            $sth->execute();
         }
 
         /**
@@ -177,9 +205,10 @@
          * @return int 若存在則回傳該網址之ID，若不存在則回傳 0
          */
         private function checkURL($URL){
-            $query = "SELECT HASH,ID FROM URL WHERE HASH =?";
+            $query = "SELECT HASH,ID FROM URL WHERE HASH = :HASH";
             $sth = $this->db->prepare($query);
-            $sth->execute( array(md5($URL)) );
+            $sth->bindParam(':HASH', md5($URL));
+            $sth->execute();
             foreach($sth as $row){
                 if($row['HASH'] == md5($URL) ){
                     return $row['ID'];
@@ -197,9 +226,11 @@
          * @return void
          */
         public function deleteURL($user_id,$URL){
-            $query = "DELETE FROM List WHERE UID=? and URL_ID=?";
+            $query = "DELETE FROM List WHERE UID = :user_id and URL_ID = :URL_ID";
             $sth = $this->db->prepare($query);
-            $sth->execute( array($user_id,checkURL($URL)) );
+            $sth->bindParam(':user_id', $user_id);
+            $sth->bindParam(':URL_ID', $this->checkURL($URL));
+            $sth->execute();
         }
     }
 ?>
